Backend/playlist bis (#31)
* prisma update: add field to existing tables and created profile table

* add API to playlists to add/remove songs, and get favorite

* apply swaggers to playlist + create favorite

* protect playlist deletion; add type to artist for search purpose; add new route to artist to get the artist musics

* change controllers param to dto

---------

+++ backend/src/playlists/playlists.service.ts
@@ -6,6 +6,7 @@ import {
 import { Prisma } from '../../generated/prisma/client';
 import { PrismaService } from '../prisma/prisma.service';
 import { playlist } from '../../generated/prisma/browser';
+import { PlaylistVersionResponseDto } from './dto/playlists.dto';
 
 @Injectable()
 export class PlaylistsService {
@@ -109,8 +110,22 @@ export class PlaylistsService {
     }
   }
 
-  async delete(where: Prisma.playlistWhereUniqueInput) {
+  async delete(where: Prisma.playlistWhereUniqueInput, userId: number, isAdmin: boolean) {
     try {
+	  const playlist = await this.prisma.playlist.findUnique({
+		where,
+		select: { userId: true, isDefault: true },
+	  });
+
+	  if (!playlist) {
+		throw new NotFoundException('Playlist not found');
+	  }
+	  if (playlist.userId !== userId) {
+		throw new BadRequestException('You are not the owner of this playlist');
+	  }
+	  if (playlist.isDefault && !isAdmin) {
+		throw new BadRequestException('Default playlists cannot be deleted');
+	  }
       const deletedPlaylist = await this.prisma.playlist.delete({
         where,
       });
@@ -139,31 +154,83 @@ export class PlaylistsService {
     return updatedPlaylist.version;
   }
 
-  async addMusic(playlistId: number, songId: number): Promise<number | void> {
-    const playlist = await this.prisma.playlist.findUnique({
-      where: { id: playlistId },
-      include: { musics: true },
-    });
+  async addMusic(playlistId: number, songId: number): Promise<PlaylistVersionResponseDto | void> {
+	try {
+		const playlist = await this.prisma.playlist.findUnique({
+		where: { id: playlistId },
+		include: { musics: true },
+		});
 
-    if (!playlist) {
-      throw new NotFoundException('Playlist not found');
-    }
+		if (!playlist) {
+		throw new NotFoundException('Playlist not found');
+		}
 
-    const existingMusic = playlist.musics.find((pm) => pm.musicId === songId);
+		const existingMusic = playlist.musics.find((pm) => pm.musicId === songId);
 
-    if (existingMusic) {
-      throw new BadRequestException('Music already in playlist');
-    }
+		if (existingMusic) {
+		throw new BadRequestException('Music already in playlist');
+		}
 
-    await this.prisma.playlistMusic.create({
-      data: {
-        playlistId,
-        musicId: songId,
-        index: playlist.musics.length,
-      },
-    });
+		await this.prisma.playlistMusic.create({
+			data: {
+				playlistId,
+				musicId: songId,
+				index: playlist.musics.length,
+			},
+		});
 
-    return await this.incrementPlaylistVersion(playlistId);
+		const version =await this.incrementPlaylistVersion(playlistId);
+		return {
+			playlistId: playlistId,
+			musicId: songId,
+			version: version,
+		};
+	} catch (error) {
+		if (error instanceof Prisma.PrismaClientKnownRequestError) {
+			throw new BadRequestException('Invalid data for adding music to playlist');
+		}
+		throw error;
+	}
+  }
+
+  async removeMusic(playlistId: number, songId: number): Promise<PlaylistVersionResponseDto | void> {
+	try {
+		const playlist = await this.prisma.playlist.findUnique({
+		where: { id: playlistId },
+		include: { musics: true },
+		});
+
+			if (!playlist) {
+		throw new NotFoundException('Playlist not found');
+		}
+
+		const existingMusic = playlist.musics.find((pm) => pm.musicId === songId);
+
+		if (!existingMusic) {
+		throw new BadRequestException('Music not in playlist');
+		}
+
+		await this.prisma.playlistMusic.delete({
+		where: {
+			playlistId_musicId: {
+			playlistId,
+			musicId: songId,
+			},
+		},
+		});
+
+		const version = await this.incrementPlaylistVersion(playlistId);
+		return {
+			playlistId: playlistId,
+			musicId: songId,
+			version: version,
+		};
+	} catch (error) {
+		if (error instanceof Prisma.PrismaClientKnownRequestError) {
+			throw new BadRequestException('Invalid data for adding music to playlist');
+		}
+		throw error;
+	}
   }
 
   async canJoinPlaylist(playlistId: number, userId: number): Promise<number> {
@@ -277,5 +344,67 @@ export class PlaylistsService {
     }
 
     return await this.incrementPlaylistVersion(playlistId);
+  }
+
+  async getPersonnal(userId: number) {
+    const personnalPlaylists = await this.prisma.playlist.findMany({
+      where: {
+        OR: [
+          { userId },
+          {
+            playlistships: {
+              some: {
+                addresseeId: userId,
+                status: 'ACCEPTED',
+              },
+            },
+          },
+        ],
+      },
+	  include: {
+		musics: {
+			include: {
+					music: {
+						select: {
+							duration: true,
+						},
+					},
+				},
+			},
+		},
+	});
+
+    if (!personnalPlaylists) {
+      throw new NotFoundException('No personnal playlists found');
+    }
+
+    return personnalPlaylists.map((playlist) => ({
+		id: playlist.id,
+		title: playlist.title,
+		songs: playlist.musics.length,
+		duration: playlist.musics.reduce(
+			(sum, musics) => sum + musics.music.duration,
+			0,
+		),
+		}));
+  }
+
+  async canAccess(playlistId: number, userId: number): Promise<boolean> {
+	const playlist = await this.prisma.playlist.findUnique({
+		where: {
+		  id: playlistId,
+		OR: [
+			{ userId },
+			{playlistships: {
+				some: {
+					addresseeId: userId,
+					status: 'ACCEPTED',
+				},
+			}},
+		],
+	   },
+	  select: { userId: true },
+	});
+	return playlist?.userId === userId;
   }
 }
